LayoutTests/platform/ios-simulator/editing/pasteboard/5065605-expected.txt

EDITING DELEGATE: shouldChangeSelectedDOMRange:(null) toDOMRange:(null) affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
This tests for a bug where text copied with Select All + Copy would lose its color.
| <br>
